module.git: initial commit
This commit prototypes a `module.git` module loader. `module.git` clones
and manages a Git repository and loads a modules from a file within the
managed repository.

The revision of the repository to retrieve can be any recognized
revision; including branch names, tag names, or individual commit SHAs.

The repository reference can be anything recognized by `git clone`,
including the local filesystem.

The repository will be fetched with the given pull_frequency argument.
If pull_frequency is 0, then the repository will only be re-fetched when
the component is updated. Setting this argument is primarily useful when
pointing `module.git` at a branch.

I have published a repository at
https://github.com/rfratto/agent-modules to be used with this PR for
examples.

This configuration file can load one of the example modules from that
repository:

    logging {
            level = "debug"
    }

    module.git "add" {
            repository = "https://github.com/rfratto/agent-modules.git"
            revision   = "main"
            path       = "add/module.river"

            arguments = {
                    a = 15,
                    b = 45,
            }
    }