docs: add _Collect Prometheus metrics_ topic for Flow
Add a _Collect Prometheus metrics_ topic in the Grafana Agent Flow
documentation. This topic is intended to be used for all common tasks
around Prometheus metric collection.

For this initial commit, the only task is for configuring a
prometheus.remote_write component. More tasks will be added in follow-up
PRs to keep the review size as small as possible.