docs: add Collect OpenTelemetry data topic for Flow
Add a _Collect OpenTelemetry data_ topic in the Grafana Agent Flow
documentation. This topic is intended to be used for all common tasks
around OpenTelemetry data collection and processing.

For this initial commit, the only task is for configuring an exporter
(otelcol.exporter.otlp). More tasks will be added in follow-up
PRs to keep the review size as small as possible.