module.string: fix issue with nested controller API
This commit fixes a bug where API endpoints for nested controllers did
not propagate correctly. The controller handler assumed that the full
HTTP path was being used rather than the truncated path given to
component handlers.

This also fixes an issue in the UI where the URL for deeply nested
modules was calculated incorrectly.

Fixes #3294