cmd/grafana-agent-flow: add new Flow-only command
cmd/grafana-agent-flow is a Flow-only copy of cmd/grafana-agent. It is
intended to be an intermediate build asset on systems which want to
install static mode alongside Flow mode.

cmd/grafana-agent-flow is not expected to be published as a release
asset separately, and will only be included in Flow-mode system
packages.