flow: prototype retrieving flags from config file
This commit adds a new `--flags-config-file` flag to the `run` command.
When this flag is passed and given a file, the file is parsed as a River
config representing the flags to pass to run.

For example, `--server.http.listen-addr=127.0.0.1:8080
--disable-reporting` is represented as:

    server {
      http {
        listen_addr = "127.0.0.1:8080"
      }
    }

    disable_reporting = true

This can be useful for systems where it is not easy to change options
passed to the service (like macOS; see grafana/homebrew-grafana#47).