cmd/grafana-agent-service: add serviceManager abstraction
This commit introduces a cmd/grafana-agent-service command, which will
be used for managing a Grafana Agent Flow process running on Windows.
cmd/grafana-agent-service is responsible for implementing the Windows
Service interface, and will run the Grafana Agent binary.

Additionally, the serviceManager abstraction is introduced, which can
manage a child process until a context is canceled. If the child process
exits, the serviceManager will restart it with an exponential backoff.
The exponential backoff resets as long as the process has been alive for
twice the exponential backoff period.

Closes #3243.