prometheus.remote_write: fix issue where unnecessary directory was used
prometheus.remote_write was originally creating a folder called
`wal/COMPONENT_ID` relative to the data path given to the component.

Because the data path given to the component is already unique, AND
because the Prometheus storage code creates a WAL directory, this means
that the WAL was being stored at
`DATA_PATH/COMPONENT_ID/wal/COMPONENT_ID/wal`. This issue compounds
itself when using modules, as each module adds an additional unnecessary
directory to the path.

This commit changes the WAL to be stored at
`DATA_PATH/COMPONENT_ID/wal`.