docs: split About Grafana Agent sections
The "About Grafana Agent" section mostly covered different topics for
static mode, with some content specific to Grafana Agent Operator.

This commit moves those tables to the home document for static mode and
the operator respectively.

The "About Grafana Agent" document will be replaced in a future commit
to talk about the different modes of Grafana Agent and the project as a
whole.