Introduce `component.LeastHealthy` utility function and fix bug in module.file (#3545)
* component: add LeastHealthy utility function

`LeastHealthy` takes a set of Healths and returns the one which is least
Healthy. Healths are priotized by type, then by time, followed by
argument order.

The unhealthiness is determined in descending order of:

1. HealthTypeExited (least healthy)
2. HealthTypeUnhealthy
3. HealthTypeUnknown
4. HealthTypeHealthy

* flow: use component.LeastHealthy

Replace the existing least-healthy determination with the new utility
method.

* module.file: fix issue where health of local.file component gets hidden

This fixes an issue where the health reported by the managed
`local.file` component was hidden from the overall health reporting of
the `module.file` component.