misc: update release note to link to new documentation (#3622)
The links to documentation will change with the v0.33 release, so this
commit updates the release note to use the proper links.

This also takes the opportunity to offload all of the instructions to
the documentation rather than duplicating instructions between the
documentation and the release notes.

Co-authored-by: Mischa Thompson <mischabear@gmail.com>
(cherry picked from commit e09313c82a77a805fc6d9a057a2ed61f0688f0de)