loki.source.heroku: use an unchecked collector for server metrics
Every time an HTTP server for `loki.source.heroku` is created, it
creates a new weaveworks/common server, and registers the same set of
metrics to the provided collector.

By default, these are checked metrics, which causes the agent-wide
`/metrics` endpoint to stop working with an error:

    An error has occurred while serving metrics:

    4 error(s) occurred:
    * collected metric "loki_source_heroku_drain_target_tcp_connections_limit" { label:<name:"component_id" value:"loki.source.heroku.repro" > label:<name:"protocol" value:"http" > gauge:<value:0 > } was collected before with the same name and label values
    * collected metric "loki_source_heroku_drain_target_tcp_connections_limit" { label:<name:"component_id" value:"loki.source.heroku.repro" > label:<name:"protocol" value:"grpc" > gauge:<value:0 > } was collected before with the same name and label values
    * collected metric "loki_source_heroku_drain_target_tcp_connections" { label:<name:"component_id" value:"loki.source.heroku.repro" > label:<name:"protocol" value:"http" > gauge:<value:0 > } was collected before with the same name and label values
    * collected metric "loki_source_heroku_drain_target_tcp_connections" { label:<name:"component_id" value:"loki.source.heroku.repro" > label:<name:"protocol" value:"grpc" > gauge:<value:0 > } was collected before with the same name and label values

This commit adds a new utility method to create an unchecked
prometheus.Collector implementation, and then creates a new
prometheus.Registry to store server metrics any time a new server is
created. The prometheus.Registry instance for the server is then passed
to the unchecked prometheus.Collector for handling.

Closes #3646.