loki.source.file: fix goroutine leak
Previously, one loki.EntryHandler was created per target, but only the
most recent loki.EntryHandler was tracked in memory. This caused all the
other loki.EntryHandlers to permanently leak whenever the component was
reloaded.

PR #3036 attempted to initially address the leak, but we missed that
handlers were being created in a loop in the case of loki.source.file.

Closes #3728.