flow: only use component health when component implements component.HealthComponent (#3740)
* Revert "Default back to healthy when a component does not implement component.HealthComponent (#3558)"

This reverts commit b2f8992ce0c8a7e9de53c871e64774c2ae8f8d8e. The commit
introduced a bug where the "default health" message was always returned
for healthy components, overriding the message informing the user when
the last time the component evaluated was.

* flow: only use component health if component exposes health

This commit offers an alernative approach to #3558, where the overall
health of a component only considers the component implementation of
health if that component implements the component.HealthComponent
interface.

Without this commit, the zero value of component.Health was always
returned, considering the health of every component as being "unknown."

(cherry picked from commit ff2ca95a9234f1ea49f61135ed5eaf49bd586494)