cluster: honor deadline when opening TCP connection (#3764)
Although synchronous writes to peers have a timeout, this timeout only
affected the HTTP/2 request and not the net.Dial call. This commit
attempts to honor the deadline when establishing a connection to a peer,
otherwise it falls back to a default timeout of 30s.

(cherry picked from commit f4c0e0666060d3e55d7d2587e27f1db822bd9155)