river-jsonnet: change where index gets injected to end
Existing usages of river-jsonnet may check for std.startsWith assuming
the old internal format. This change makes the prefix of internal format
consistent with what it used to be.

A longer-term solution would be to add a utility method to deconstruct
these lines rather than relying on an internal representation.