docs: fix macOS log locations for Flow
Previously, an instance of the grafana/grafana-agent-flow Homebrew
formula installed on macOS would generate logs in the same location as
the static mode grafana-agent formula.

The documentation here was incorrect, saying that the stderr file would
write to a Flow-specific file.

grafana/homebrew-grafana#52 fixes the formula to log files in the
proper, Flow-specific, place, and this PR updates the documentation to
use the correct file names (assuming grafana/homebrew-grafana#52 is
merged).