clusterer: use a 512-token ring
Use a 512-token ring to achieve better distribution. Because a 512-token
ring only needs 12MB for 1,000 nodes (compared to 6MB for 1,000 nodes in
a 256-token ring), its overhead is marginal enough to be worth it for
the improved distribution.

However, this change means that any existing cluster being upgraded to
the new version will suffer from incorrect or missing hashing
assignments until all nodes are using 512 tokens.

This is the type of change we shouldn't make often; I think it's only
feasible now because clustering is still experimental. Future changes
would likely need to be opt-in via a flag.