river/encoding/riverjson: refactor river/encoding
This commit refactors river/encoding as river/encoding/riverjson, which
aligns more closely with the encoding implementation used by
river/token/builder.

Aligning with the implementation more closely allows bug fixes to more
easily be ported across the two implementations, and allows #2811 to be
implemented more effectively in the same style as #4057.