pkg/flow: add APIs to list/get component details
This commit adds a Flow.GetComponent and Flow.ListComponents methods to
allow retrieving detailed information on running components.

Part of the API is unimplemented; there is the intent that these
endpoints will be able to return all components, including components
inside of modules, but this is left unimplemented while the work in
grafana/agent#4027 is ongoing. For now, the controller is limited to
returning its own components.