misc: remove dependency on k3d
This commit removes the dependency on k3d, replacing it with the
existing dependency on testcontainers-go and the k3s module.

k3d is a very large module to have a dependency just for tests; it makes
makes updating other dependencies more difficult. Switching it with k3d
and testcontainers-go ensures that our direct dependencies have less
overlap so updating things is easier.