Flow: add new methods to get and list components (#4097)
* pkg/flow: use consistent receiver names for Flow

This commit changes the receiver names to always be `f` rather than a
mixture of `c` and `f`.

* pkg/flow: add APIs to list/get component details

This commit adds a Flow.GetComponent and Flow.ListComponents methods to
allow retrieving detailed information on running components.

Part of the API is unimplemented; there is the intent that these
endpoints will be able to return all components, including components
inside of modules, but this is left unimplemented while the work in
grafana/agent#4027 is ongoing. For now, the controller is limited to
returning its own components.

* web/api: change listComponentHandler to getComponentHandler

listComponentsHandler and listComponentHandler were named too similarly,
making it hard to differentiate between the two.

This commit renames the get method so the method pair is now
listComponentsHandler and getComponentHandler.

* web/api: use new Flow APIs for web API

Change the implementation of web/api to use the new Flow
ListComponents/GetComponent methods to power the data returned via the
API.

* module.string: allow tests to run on macOS as well as Linux

* module.string: use new Flow APIs in tests

* cmd/internal/flowmode: use new Flow APIs for getting enabled components

* pkg/flow: remove unused code

Remove unused code from previous implementation of getting component
information for the HTTP API.

* web/api: move provider definition to component package

This is a bit of a code smell, but it's the best place to handle this
for now. It should be easier to work around it in the future with the
services proposal without having to have too many type definitions in a
single package with no logic.

* web/api/apitypes: remove unused package

The web/api/apitypes package is now unused and can be removed; it's
replaced with the detailed generic component information which is also
serializable to JSON.