prometheus.remote_write: do not fall back to upstream defaults
When the `queue_config` or `metadata_config` blocks were fully
unspecified, their values were inherited from the Prometheus package.
This is an issue in cases where we purposefully changed the values we
used.

A temporary workaround for the user is to specify empty blocks so the
correct defaults are used (`queue_config {}`, `metadata_config {}`).

Fixes #4164.