river/vm: unwrap nonsesitive OptionalSecret before performing binary ops (#4288)
This change unwraps an OptionalSecret (when IsSecret is false) to a
string for the purposes of binary operations. This allows OptionalSecret
to be compared to strings and concatenated with strings, including with
other OptionalSecret values.

This is a hack to fix the observable behavior, but the nature of the fix
means that other issues may still occur in the future if there are ever
more capsule types which support conversion to a non-capsule type. A
longer-term fix is needed once we identify what type of conversions
binary operations should do.

Fixes #4036.