Fix kubelet discovery for deleted pods
The discovery component requires a target group to be in the list of
discovered targets with an empty source when a target should be removed.
Therefore we need to cache the previous discovery results in order to
delete them in the next discovery run if the pod no longer exists.