pkg/flow: implement service.Host in Flow controller
This commit updates the Flow controller to implement the new
service.Host interface.

Note that because it is not currently possible for a service or
component to define a dependency on a service, the GetServiceConsumers
method always returns nil.

GetServiceConsumers will be updated in the future to return a non-nil
list once it is possible for consumers of a service to exist.