web/api: enable a cross-module web API
This commit creates a single cross-module web API which can serve the UI
for all modules.

Previously, these were the endpoints used by the API:

* `/api/v0/web/components`: get components for root module
* `/api/v0/web/components/{id}`: get component from root module
* `/api/v0/component/{module_id}/components`: get components for non-rooted module ID
* `/api/v0/component/{module_id}/components/{id}`: get component for non-rooted module ID

These are now replaced with three endpoints instead:

* `/api/v0/web/components`: get components for root module
* `/api/v0/web/components/{global_id}`: get a component from its global ID
* `/api/v0/web/modules/{module_id}/components`: get components for a non-root module

`global_id` and `module_id` above are permitted to have slash characters
in them.

Existing usages of the API continue to work as-is.