pkg/metrics/wal: always write series records to WAL, even on rollback (#4619)
If a new series is introduced in a storage.Appender instance, that
series should be written to the WAL once the storage.Appender is closed,
even on Rollback.

Previously, new series would only be written to the WAL when calling
Commit. However, because the series is stored in memory regardless,
subsequent calls to Commit may write samples to the WAL which reference
a series ID which that was never written.

Port of prometheus/prometheus#12592

(cherry picked from commit 3837d47dbbfff039b49c5a8272a97c52feb4b58b)