flow: enable the management of services
This change adds a new field to flow.Options, Services, which causes the
lifecycle management of services to be handled by the Flow controller.

Services are added to the DAG as nodes, and are run with the Flow
controller when the Flow controller starts. Services that declare a
dependency on another service are only evaluated after the services they
depend on have been evaluated.

GetServiceConsumers will now currently return instances of
component.Component and service.Service which depend on a service. The
set of component.Component dependants is currently empty since it is not
yet possible for a component to define a dependency on a service.

The follow items are left as follow-up work:

* Allow components to define a dependency on a service.

* Allow service data to be exposed to components which depend on that
  service.

* Allow services to propagate to loaded modules.

* Migrate existing services to be managed by the Flow controller.

Related to #4253.