flow: enable managing services (#4664)
* flow/internal/controller: rename files holding DAG node implementations

Rename all files holding DAG node implementations to start with node_ so
they're easier to locate.

* flow/internal/controller: add ServiceNode

ServiceNode is the representation of a Flow service in the DAG.

Note that it is not currently wired up anywhere.

* flow/internal/controller: make NewLoader more flexible

Create a new type, LoaderOptions, which passes options to a Loader, to
make extending Loader with new options more flexible.

LoaderOptions currently encapsulates ComponentGlobals, but will be
expanded in the future to contain other relevant options (such as
services to load in the DAG).

* flow/internal/controller: add service nodes to graph

Add service nodes to the graph when the graph is being constructed. If a
service declares a dependency on another service, it forms a DAG edge.

* flow/internal/controller: expose ServiceNodes to caller

This change stores and exposes ServiceNodes after calls to Apply.

* flow: enable the management of services

This change adds a new field to flow.Options, Services, which causes the
lifecycle management of services to be handled by the Flow controller.

Services are added to the DAG as nodes, and are run with the Flow
controller when the Flow controller starts. Services that declare a
dependency on another service are only evaluated after the services they
depend on have been evaluated.

GetServiceConsumers will now currently return instances of
component.Component and service.Service which depend on a service. The
set of component.Component dependants is currently empty since it is not
yet possible for a component to define a dependency on a service.

The follow items are left as follow-up work:

* Allow components to define a dependency on a service.

* Allow service data to be exposed to components which depend on that
  service.

* Allow services to propagate to loaded modules.

* Migrate existing services to be managed by the Flow controller.

Related to #4253.