component: allow components to depend on a service
This commit allows a component to define a dependency on a service. When
wiring dependencies in the graph, the registration of a component is
used to create the dependency.

This also means that Flow.GetServiceConsumers will now implicitly return
instances of [component.Component] which declare a dependency on the
service.

Related to #4253.