component: expose service data to components
This commit updates the Flow controller to expose service data to
components.

Components are only allowed to access service data for services they
have listed as an explicit dependency. This ensures that services cannot
access data for a service they depend on until that service has been
evaluated.

Related to #4253.