component: allow components to consume data for services  (#4678)
* component: allow components to depend on a service

This commit allows a component to define a dependency on a service. When
wiring dependencies in the graph, the registration of a component is
used to create the dependency.

This also means that Flow.GetServiceConsumers will now implicitly return
instances of [component.Component] which declare a dependency on the
service.

Related to #4253.

* component: expose service data to components

This commit updates the Flow controller to expose service data to
components.

Components are only allowed to access service data for services they
have listed as an explicit dependency. This ensures that services cannot
access data for a service they depend on until that service has been
evaluated.

Related to #4253.

* flow: add tests for components using services

---------

Co-authored-by: Paschalis Tsilias <tpaschalis@users.noreply.github.com>