flow: fix bug where GetServiceConsumers didn't search through loaded modules
GetServiceConsumers will be used by services to quickly retrieve a list
of components and services which depend on a service for service-related
operations.

For example, the clustering service will use GetServiceConsumers to know
which components to notify of a cluster update.

GetServiceConsumers needs to return dependant components throughout all
modules, otherwise the clustering service will be unable to notify all
appropriate components when the clustering state changes.