flow: change clustering interface to be a call to NotifyClusterChange
Previously, components would opt in to clustering notifications through
a call to ClusterUpdatesRegistration. Then, if that method returned
true, they would be re-evaluated.

In preparation for the move to services, the clustering service would
not be able to function the same way, as services are not permitted to
trigger a re-evaluation of a component.

The new interface is a NotifyClusterChange method, which allows
components to opt-in to whether they want to ignore that signal or not.

This is not reflected in the CHANGELOG as it is a no-op for users.