service/cluster: introduce cluster service
The cluster service is a replacement for pkg/cluster that implements the
github.com/grafana/agent/service.Service API.

It has the following changes compared to the previous implementation:

* service/cluster always starts a ckit.Node for consistency around the
  API. When clustering is disabled, service/cluster never joins any
  peers, and never exposes its cluster HTTP handlers to peers in the
  cluster.

* Much of the behavior previously provided by pkg/cluster is now
  deferred to the caller:

  * The caller is expected to determine their own advertise address,
    including defaulting to a specific interface.

  * The caller is expected to provide their own function to discover
    peers, even if that function returns a static list.

* The new implementation, as a Flow service, directly informs components
  about changes to the cluster state instead of having the Flow
  controller inform components.

While the new implementation puts more work on the caller, its API is
simplified. Mutually exclusive mechanisms of the previous API (such as
configuring JoinPeers and DiscoverPeers) have been simplified into a
single, unified mechanism.

While this commit introduces the new service, existing code is not
updated the use the new service yet.