all: replace usages of pkg/cluster.Clusterer with pkg/cluster.Node
This changes all usages of pkg/cluster.Clusterer to use an interface
instead. Usage of the interface will permit changing out the
implementation, particularly with something from service/cluster.

This is a temporary change for the purposes of atomic commits. As such,
the Clusterer field is not renamed to the more appropriate name of
"Node."