service/cluster: migrate pkg/cluster to a Flow-native service (#4870)
* service/http: allow dependant services to register custom handlers

This updates the HTTP service to allow downstream services to register
custom handlers.

This will replace ths custom logic for wiring methods for the HTTP
service, which currently requires the clustering service to be
explicitly passed to the HTTP service.

* service/http: move UI-specific routes to a dedicated UI service.

The UI service depends on both the HTTP and clustering services to power
the API, which means it should be brought out to a dedicated package to
avoid a cyclic dependency.

* service/cluster: introduce cluster service

The cluster service is a replacement for pkg/cluster that implements the
github.com/grafana/agent/service.Service API.

It has the following changes compared to the previous implementation:

* service/cluster always starts a ckit.Node for consistency around the
  API. When clustering is disabled, service/cluster never joins any
  peers, and never exposes its cluster HTTP handlers to peers in the
  cluster.

* Much of the behavior previously provided by pkg/cluster is now
  deferred to the caller:

  * The caller is expected to determine their own advertise address,
    including defaulting to a specific interface.

  * The caller is expected to provide their own function to discover
    peers, even if that function returns a static list.

* The new implementation, as a Flow service, directly informs components
  about changes to the cluster state instead of having the Flow
  controller inform components.

While the new implementation puts more work on the caller, its API is
simplified. Mutually exclusive mechanisms of the previous API (such as
configuring JoinPeers and DiscoverPeers) have been simplified into a
single, unified mechanism.

While this commit introduces the new service, existing code is not
updated the use the new service yet.

* flowmode: add utilities to build a cluster.Service implementation

The implementation of service/cluster shifts more responsibility to the
caller than was previously required with pkg/cluster. This commit adds
helper utilities to implement that responsibility to retain the same
behavior with the new service.

* all: replace usages of pkg/cluster.Clusterer with pkg/cluster.Node

This changes all usages of pkg/cluster.Clusterer to use an interface
instead. Usage of the interface will permit changing out the
implementation, particularly with something from service/cluster.

This is a temporary change for the purposes of atomic commits. As such,
the Clusterer field is not renamed to the more appropriate name of
"Node."

* service/cluster: temporarily match pkg/cluster.Node interface

This is a temporary commit for service/cluster.Cluster to match the
pkg/cluster.Node interface.

The purpose of this commit is to atomically swap out the implementation
of pkg/cluster with service/cluster without needing to remove the
hard-coded usage of the interface yet.

* flowmode: Replace usage of pkg/cluster with service/cluster.

* flow: use GetServiceData API and remove hard-wired cluster usage

* flow: remove remaining references to pkg/cluster

* service/cluster: remove unused methods

Remove methods which were temporarily added to support an atomic commits
of switching pkg/cluster usage with service/cluster.

* misc: remove pkg/cluster

pkg/cluster is now considered dead code in favor of the flow-specific
service/cluster API.

* component: remove reference to old ClusteredComponent interface

The ClusteredComponent interface has been replaced with the interface
definition in cluster.Component.