integrations/eventhandler: remove double-logging of events (#5100)
This removes the info-level log line indicating that an event is logged.
When agent logs are being collected, this makes the event appear as if
it was double-collected, as the event appears both in the logs sent
directly to Loki and in the agent logs.

(cherry picked from commit 276355ee0d54a81e12ed186f949645eb9d37ff9a)
Signed-off-by: erikbaranowski <39704712+erikbaranowski@users.noreply.github.com>