flow/logging: use log/slog internally for log handling (#5107)
This commit updates the Flow logger to expose a log/slog.Handler. The
slog.Handler implementation always reflects the current logging
settings, even when calling WithAttrs or WithGroup.

As of this commit, flow/logging.Logger continues to implement the
log.Logger interface, but this will be removed once all references to
go-kit/log are removed.

Note that with this change, slog always puts the log message at the end
of the set of builtins (timestamp, level, source, message). Previously,
there may have been other fields which appeared before the message.