otelcol: fix retry mechanism
The v0.80 upgrade introduced two new settings to the common retry
mechanism shared between components. One of these new settings,
`multiplier`, specifies the rate at which a backoff is applied.

We did not convert this new field after upgrading, leading the
multiplier to be set to the Go default (zero). After the first retry,
the backoff period would be multiplied by zero for each subsequent
retry, causing retries to happen in a hot loop.