Add faro.receiver component (#5314)
This commit adds a component for faro.reciever, an equivalent to the app_agent_receiver integration from static mode. 

This is not 100% a straight port from the existing integration. I have done some minor refactoring to make it possible for this component to be dynamically updateable at runtime. However, the implementation remains mostly the same. 

During my porting, I have noticed bugs in integration, particularly around sourcemaps. I have not fixed these, but I have added TODO comments in the code for them to be resolved down the line. 

Unfortunately, the resulting code is quite complex. Future changes to refactor it to make it easier to read are warranted, but it would likely require more code and would drift further from the original integration. In the spirit of this being a "port," such changes are not included here. 

Supersedes grafana/agent#5243 
Closes grafana/agent#2320

Co-authored-by: CRA Keishi Kawada <keishi.kawada@craftsman-software.com>
Co-authored-by: mattdurham <mattdurham@ppog.org>
Co-authored-by: Clayton Cornell <131809008+clayton-cornell@users.noreply.github.com>
Co-authored-by: Paschalis Tsilias <tpaschalis@users.noreply.github.com>