ci: automate starting new release cycles when release branch is cut
This commit introduces new automation to automatically bump the version
in the main branch at the point a release branch is created.

For example, if branch release-v0.37 is pushed, this action will trigger
and perform the following:

* Clone the repo off of the main branch.
* Determine the release cycle after release-v0.37 (v0.38.0).
* Update the version file and regenerate versioned templates.
* Open a PR against the main branch.

This automation reduces the overhead of PRs like grafana/agent#5384,
where it'll no longer be necessary to manually update the version file
at the start of a new release cycle.

This action has been tested extensively on
<https://github.com/rfratto/agent> and was shown to work as expected.