loki.source.kubernetes: improve detection of rolled log files
Versions of Kubernetes that do not contain kubernetes/kubernetes#115702
will fail to detect rolled log files, causing the API to stop sending
logs to the agent for processing.

To work around this, this commit intorduces a rolling average calculator
to determine the average delta between log entries per target. If 3x the
normal delta time has elapsed since the last entry, the tailer is
restarted.

False positives here are acceptable, but false negatives mean that log
lines may not appear for an extended period of time until the rolling
detection succeeds.

Closes #5040

Co-authored-by: Edward Welch <edward.welch@grafana.com>